requirements.txt
elasticsearch-dsl>=2.0.0,<3.0.0
pythondialog
